+++ futures-channel/src/mpsc/mod.rs
@@ -78,7 +78,7 @@ use std::usize;
 
 use mpsc::queue::{Queue, PopResult};
 use task::{self, Task};
-use {Async, AsyncSink, StartSend, Poll, Stream};
+use {Async, Poll, Stream};
 
 mod queue;
 
@@ -414,19 +414,21 @@ impl<T> Sender<T> {
     /// Attempt to start sending a message on the channel.
     ///
     /// If there is not room on the channel, this function will return
-    /// `Ok(AsyncSink::Pending(msg))`, and the current `Task` will be
+    /// `Ok(Err(msg))`, and the current `Task` will be
     /// awoken when the channel is ready to receive more messages.
-    pub fn start_send(&mut self, msg: T) -> StartSend<T, SendError<T>> {
+    ///
+    /// On successful completion, this function will return `Ok(Ok(()))`.
+    pub fn start_send(&mut self, msg: T) -> Result<Result<(), T>, SendError<T>> {
         // If the sender is currently blocked, reject the message before doing
         // any work.
         if !self.poll_unparked(true).is_ready() {
-            return Ok(AsyncSink::Pending(msg));
+            return Ok(Err(msg));
         }
 
         // The channel has capacity to accept the message, so send it.
         self.do_send(Some(msg), true)?;
 
-        Ok(AsyncSink::Ready)
+        Ok(Ok(()))
     }
 
     // Do the send without failing
@@ -649,9 +651,12 @@ impl<T> UnboundedSender<T> {
     /// Attempt to start sending a message on the channel.
     ///
     /// If there is not room on the channel, this function will return
-    /// `Ok(AsyncSink::Pending(msg))`, and the current `Task` will be
+    /// `Ok(Err(msg))`, and the current `Task` will be
     /// awoken when the channel is ready to receive more messages.
-    pub fn start_send(&mut self, msg: T) -> StartSend<T, SendError<T>> {
+    ///
+    /// On a successful `start_send`, this function will return
+    /// `Ok(Ok(())`.
+    pub fn start_send(&mut self, msg: T) -> Result<Result<(), T>, SendError<T>> {
         self.0.start_send(msg)
     }
