+++ futures-sink/src/channel_impls.rs
@@ -1,12 +1,19 @@
 use {Async, Sink, AsyncSink, StartSend, Poll};
 use futures_channel::mpsc::{Sender, SendError, UnboundedSender};
 
+fn res_to_async_sink<T>(res: Result<(), T>) -> AsyncSink<T> {
+    match res {
+        Ok(()) => AsyncSink::Ready,
+        Err(x) => AsyncSink::Pending(x),
+    }
+}
+
 impl<T> Sink for Sender<T> {
     type SinkItem = T;
     type SinkError = SendError<T>;
 
     fn start_send(&mut self, msg: T) -> StartSend<T, SendError<T>> {
-        self.start_send(msg)
+        self.start_send(msg).map(res_to_async_sink)
     }
 
     fn poll_complete(&mut self) -> Poll<(), SendError<T>> {
@@ -23,7 +30,7 @@ impl<T> Sink for UnboundedSender<T> {
     type SinkError = SendError<T>;
 
     fn start_send(&mut self, msg: T) -> StartSend<T, SendError<T>> {
-        self.start_send(msg)
+        self.start_send(msg).map(res_to_async_sink)
     }
 
     fn poll_complete(&mut self) -> Poll<(), SendError<T>> {
